requirements.txt
Django==3.2.18
djangorestframework==3.13.1
django-rest-swagger==2.2.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
celery==5.2.3
redis==4.1.0
drf-yasg==1.20.0